Call-Agent/callagent/requirements.txt
python-dotenv==1.0.0
loguru==0.7.2
twilio==8.11.0
langgraph==0.2.45
langchain-core==0.3.15
flask==3.0.0
flask-cors==4.0.0
